backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.4
celery==5.3.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
slowapi==0.1.9
prometheus-fastapi-instrumentator==6.1.0
structlog==24.1.0
httpx==0.27.0
python-dotenv==1.0.1
email-validator==2.1.1
pytz==2024.1
